+++ libraries/test_session_current/src/main/java/androidx/media3/session/RemoteMediaSession.java
@@ -447,6 +447,10 @@ public class RemoteMediaSession {
       binder.notifyTimelineChanged(sessionId, reason);
     }
 
+    public void notifyTimelineChangedWithMediaItemTransition() throws RemoteException {
+      binder.notifyTimelineChangedWithMediaItemTransition(sessionId);
+    }
+
     public void notifyPlaylistMetadataChanged() throws RemoteException {
       binder.notifyPlaylistMetadataChanged(sessionId);
     }